requirements.txt
numpy<2
aiofiles==23.2.1
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
pydantic==2.7.1
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.3.0
PyMuPDF==1.24.2
pytest==8.2.0
httpx==0.27.0